% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckServerRequirements.R
\name{CheckServerRequirements}
\alias{CheckServerRequirements}
\title{CheckServerRequirements}
\usage{
CheckServerRequirements(
  ServerSpecifications = NULL,
  RequiredPackages = c("dsBase", "dsCCPhos", "dsTidyverse"),
  RequiredFunctions = c(aggregate = "GetReportingObjectDS", assign = "AugmentDataDS",
    assign = "CurateDataDS", assign = "ExtractFromListDS", assign = "DrawSampleDS"),
  RequiredOpalTableNames = dsCCPhosClient::Meta.Tables$TableName.Raw,
  DSConnections = NULL
)
}
\arguments{
\item{ServerSpecifications}{\code{data.frame} - Same \code{data.frame} used for login. Used here only for acquisition of server-specific project names (in case they are differing). - Default: \code{NULL} for virtual project}

\item{RequiredPackages}{A \code{character vector} naming required packages}

\item{RequiredFunctions}{A named \code{character vector} containing names of required functions. Their type ('aggregate' or 'assign') is defined by the correspondent element names.}

\item{RequiredOpalTableNames}{\code{character vector} - The expected names of the Opal data base tables.}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of \code{data.frames} containing gathered info and messages
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Check if technical requirements are met on every participating server.
}
\author{
Bastian Reiter
}
