% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToVirtualCCP.R
\name{ConnectToVirtualCCP}
\alias{ConnectToVirtualCCP}
\title{ConnectToVirtualCCP}
\usage{
ConnectToVirtualCCP(
  CCPTestData,
  NumberOfServers = 1,
  NumberOfPatientsPerServer = NULL,
  AddedDsPackages = NULL,
  Resources = NULL,
  WorkingDirectory = file.path(tempdir(), ".dslite")
)
}
\arguments{
\item{CCPTestData}{\code{Named list} of \code{data.frames} - CCP test data}

\item{NumberOfServers}{\code{integer} - The number of virtual servers to install}

\item{NumberOfPatientsPerServer}{\code{integer} - Optional value to restrict size of data set for faster testing - Default: NULL}

\item{AddedDsPackages}{\code{character vector} - Server-side DataSHIELD packages to be added to default (dsBase, dsCCPhos) - Default: NULL}

\item{Resources}{\code{Named list} of \code{resourcer::Resource} objects - Default: NULL}

\item{WorkingDirectory}{\code{string} - Optional custom working directory for virtual servers - Default: Hidden folder in R session's temporary directory (see \code{?DSLite::newDSLiteServer()})}
}
\value{
A \code{list} of \code{DSConnection}-objects
}
\description{
Sets up a virtual DataSHIELD infrastructure that enables trying out real dsCCPhos functionality on test data.
}
\author{
Bastian Reiter
}
