% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.CurateData.R
\name{ds.CurateData}
\alias{ds.CurateData}
\title{ds.CurateData}
\usage{
ds.CurateData(
  RawDataSetName = "RawDataSet",
  OutputName = "CurationOutput",
  Settings = NULL,
  RunAssignmentChecks = TRUE,
  UnpackCuratedDataSet = TRUE,
  DSConnections = NULL
)
}
\arguments{
\item{RawDataSetName}{\code{character} - Name of Raw Data Set object (list) on server - Default: 'RawDataSet'}

\item{OutputName}{\code{character} - Name of output object to be assigned on server - Default: 'CurationOutput'}

\item{Settings}{\code{list} - Settings passed to function
\itemize{  \item \emph{DataHarmonization} - \code{list}
\itemize{ \item Run \code{logical} - Whether or not to perform data harmonization - Default: \code{TRUE}
\item Methods \code{data.frame} - Default: \code{dsCCPhos::Set.DataHarmonizationMethods}
\item TransformativeExpressions \code{data.frame} - Default: \code{dsCCPhos::Set.TransformativeExpressions}
\item TransformativeExpressions.Profile \code{string} - Profile used in \emph{TransformativeExpressions} - Default: 'Default'
\item Dictionary \code{data.frame} - Default: \code{dsCCPhos::Meta_Dictionary}
\item Dictionary.Profile \code{string} - Profile used in \emph{Dictionary} - Default: 'Default'
\item FuzzyStringMatching \code{data.frame} - Default: \code{dsCCPhos::Meta_FuzzyStringMatching}
\item FuzzyStringMatching.Profile \code{string} - Profile used in \emph{FuzzyStringMatching} - Default: 'Default'}
\item \emph{FeatureObligations} - \code{list}
\itemize{ \item RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_FeatureObligations}
\item RuleSet.Profile \code{string} - Profile name defining strict and trans-feature rules for obligatory feature content. Profile name must be stated in \code{FeatureObligations$RuleSet} - Default: 'Default'}
\item \emph{FeatureTracking} - \code{list}
\itemize{ \item RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_FeatureTracking}
\item RuleSet.Profile \code{string} - Profile name defining which features should be tracked/monitored during curation process. Profile name must be stated in \code{FeatureTracking$RuleSet} - Default: 'Default'}
\item \emph{TableCleaning} - \code{list}
\itemize{ \item Run \code{logical} - Whether or not to perform table cleaning (removal of redundant and ineligible entries) - Default: \code{TRUE}}
\item \emph{TableNormalization} - \code{list}
\itemize{ \item Run \code{logical} - Whether or not to perform table normalization - Default: \code{TRUE}
\item RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_TableNormalization}
\item RuleSet.Profile \code{string} - Profile name defining rule set to be used for table normalization. Profile name must be stated in \code{TableNormalization$RuleSet} - Default: 'Default'}}}

\item{RunAssignmentChecks}{\code{logical} Indicating whether assignment checks should be performed or omitted for reduced execution time - Default: \code{TRUE}}

\item{UnpackCuratedDataSet}{\code{logical} indicating whether the Curated Data Set \code{list} should be unpacked so that tables \code{data.frames} are directly accessible - Default: \code{TRUE}}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
\code{list} of following objects:
\itemize{\item 'Messages' - Info messages concerning completion of \code{CurateDataDS()} and assignment of the following objects on server:
\itemize{\item CurationOutput (\code{list})
\itemize{ \item CuratedDataSet \code{list}
\itemize{ \item BioSampling
\item Diagnosis
\item DiseaseStatus
\item GeneralCondition
\item Histology
\item Metastasis
\item MolecularDiagnostics
\item OtherClassification
\item Patient
\item RadiationTherapy
\item Staging
\item Surgery
\item SystemicTherapy
\item TherapyRecommendation}
\item CurationReport \code{list}
\itemize{\item EntryCounts \code{tibble}
\itemize{ \item Table
\item InitialCount
\item ExcludedPrimary
\item AfterPrimaryExclusion
\item ExcludedSecondary
\item AfterSecondaryExclusion
\item ExcludedSecondaryRedundancy
\item AfterSecondaryRedundancyExclusion}
\item Transformation (list of lists)
\itemize{ \item Monitors
\item EligibilityOverviews
\item ValueSetOverviews}}
\item CurationMessages \code{list}}}
\item 'CurationCompletionCheck'}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Triggers transformation of Raw Data Set (RDS) into Curated Data Set (CDS) on servers.
}
\details{
Linked to server-side ASSIGN methods \code{CurateDataDS()} and \code{ExtractFromListDS()}
}
\author{
Bastian Reiter
}
