% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.GetRDSValidationReport.R
\name{ds.GetRDSValidationReport}
\alias{ds.GetRDSValidationReport}
\title{ds.GetRDSValidationReport}
\usage{
ds.GetRDSValidationReport(RawDataSetName = "RawDataSet", DSConnections = NULL)
}
\arguments{
\item{RawDataSetName}{\code{string} - Name of 'RawDataSet' object on server - Default: 'RawDataSet'}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of \code{tibbles} containing output of validation
}
\description{
Triggers validation of Raw Data Set on server and requests a report.
}
\details{
Linked to server-side \code{AGGREGATE} method \code{GetRDSValidationReportDS()}
}
\author{
Bastian Reiter
}
