% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.P21.AssessComorbidity.R
\name{ds.P21.AssessComorbidity}
\alias{ds.P21.AssessComorbidity}
\title{ds.P21.AssessComorbidity}
\usage{
ds.P21.AssessComorbidity(
  DiagnosisData,
  DiagnosticCodeFeature,
  IDFeature,
  IgnoredCategories = NULL,
  ReturnScoreValueOnly = FALSE,
  ScoreFeatureName = "ComorbidityScore",
  Arg.map = "charlson_icd10_quan",
  Arg.assign0 = TRUE,
  Arg.weights = "quan",
  OutputName = "P21.Comorbidity",
  DSConnections = NULL,
  DS.async = dsFredaClient::Set.DSSettings$DS.async
)
}
\arguments{
\item{DiagnosisData}{\code{string} - Name of the server-side \code{data.frame} holding diagnostic code data and a grouping identifier feature}

\item{DiagnosticCodeFeature}{\code{string} - Name of the feature containing diagnostic codes}

\item{IDFeature}{\code{string} - Name of the feature containing patient- or case-specific identifier (or any other feature grouping the diagnostic codes)}

\item{IgnoredCategories}{Optional \code{character vector} - Containing morbidity categories that should be ignored in score calculation. See documentation on \code{comorbidity::comorbidity()} for reference.}

\item{ReturnScoreValueOnly}{\code{logical} - Whether the assigned \code{data.frame} should only contain the feature with comorbidity score values and not the category-specific features}

\item{ScoreFeatureName}{\code{string} - The name of the feature that holds the comorbidity score values in the assigned \code{data.frame}}

\item{Arg.map}{\code{string} - The value for argument 'map' in \code{comorbidity::comorbidity()}}

\item{Arg.assign0}{\code{logical} - The value for argument 'assign0' in \code{comorbidity::comorbidity()}}

\item{Arg.weights}{\code{string} - The value for argument 'weights' in \code{comorbidity::score()}}

\item{OutputName}{\code{string} - Assigned name of the server-side returned \code{data.frame}}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}

\item{DS.async}{\code{logical} - Value of argument 'async' in \code{DSI::datashield.assign()} / \code{DSI::datashield.aggregate()} - Default: \code{dsFredaClient::Set.DSSettings$DS.async}}
}
\value{
A \code{list} of messages about object assignment for monitoring purposes
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Using functionality in package \code{comorbidity}, map ICD codes to morbidity categories and calculate a comorbidity score.
}
\details{
Linked to server-side ASSIGN method \code{P21.AssessComorbidityDS()}
}
\author{
Bastian Reiter
}
