% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.P21.GetCohortDescription.R
\name{ds.P21.GetCohortDescription}
\alias{ds.P21.GetCohortDescription}
\title{ds.P21.GetCohortDescription}
\usage{
ds.P21.GetCohortDescription(
  DataSetName = "P21.AugmentedDataSet",
  Stage = "Augmented",
  DSConnections = NULL,
  DS.async = dsFredaClient::Set.DSSettings$DS.async
)
}
\arguments{
\item{DataSetName}{\code{string} - Name of Data Set object (list) on server, usually "RawDataSet", "CuratedDataSet" or "AugmentedDataSet" - Default: "AugmentedDataSet"}

\item{Stage}{\code{string} - Indicating the transformation stage of the described data set, one of "Raw" / "Curated" / "Augmented" - Default: "Augmented"}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}

\item{DS.async}{\code{logical} - Value of argument 'async' in \code{DSI::datashield.assign()} / \code{DSI::datashield.aggregate()} - Default: \code{dsFredaClient::Set.DSSettings$DS.async}}
}
\value{
A \code{list} containing descriptive statistics characterizing patient cohort:
\itemize{\item CohortSize
\item CohortSizeOverTime
\item AgeDistribution
\item SexDistribution}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Obtain Server-specific and aggregated descriptive characteristics about the patient cohort to be analyzed.
}
\details{
Linked to server-side \code{AGGREGATE} function \code{GetCohortDescriptionDS()}.
}
\author{
Bastian Reiter
}
