% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.P21.UnpackAugmentedDataSet.R
\name{ds.P21.UnpackAugmentedDataSet}
\alias{ds.P21.UnpackAugmentedDataSet}
\title{ds.P21.UnpackAugmentedDataSet}
\usage{
ds.P21.UnpackAugmentedDataSet(
  AugmentedDataSetName = "P21.AugmentedDataSet",
  DSConnections = NULL,
  DS.async = FALSE
)
}
\arguments{
\item{AugmentedDataSetName}{\code{string} - Name of Augmented Data Set object (\code{list}) on server - Default: 'AugmentedDataSet'}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}

\item{DS.async}{\code{logical} - Value of argument 'async' in \code{DSI::datashield.assign()} / \code{DSI::datashield.aggregate()} - Default: \code{FALSE}}
}
\value{
A \code{list} of messages about object assignment for monitoring purposes
}
\description{
Make tables within Augmented Data Set (\code{list} object) directly addressable in R server sessions.
}
\details{
Linked to server-side \code{ASSIGN} method \code{ExtractFromListDS()}
}
\author{
Bastian Reiter
}
