% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.UnpackAugmentedDataSet.R
\name{ds.UnpackAugmentedDataSet}
\alias{ds.UnpackAugmentedDataSet}
\title{ds.UnpackAugmentedDataSet}
\usage{
ds.UnpackAugmentedDataSet(
  AugmentedDataSetName = "AugmentedDataSet",
  DSConnections = NULL
)
}
\arguments{
\item{AugmentedDataSetName}{\code{string} - Name of Augmented Data Set object (\code{list}) on server - Default: 'AugmentedDataSet'}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of messages about object assignment for monitoring purposes
}
\description{
Make tables within Augmented Data Set (\code{list} object) directly addressable in R server sessions.
}
\details{
Linked to server-side \code{ASSIGN} method \code{ExtractFromListDS()}
}
\author{
Bastian Reiter
}
