% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsFredaP21Client-package.R
\docType{package}
\name{dsFredaP21Client-package}
\alias{dsFredaP21Client}
\alias{dsFredaP21Client-package}
\title{dsFredaP21Client: Enabling Data Science with P21 data}
\description{
Client-side DataSHIELD R package to enable Data Science P21 data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BastianReiter/dsFredaP21Client}
}

}
\author{
\strong{Maintainer}: Bastian Reiter \email{b.reiter@med.uni-frankfurt.de} (\href{https://orcid.org/0000-0002-1088-0502}{ORCID})

}
\keyword{internal}
