% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadRawDataSet.R
\name{LoadRawDataSet}
\alias{LoadRawDataSet}
\title{LoadRawDataSet}
\usage{
LoadRawDataSet(
  ServerSpecifications = NULL,
  RawTableNames = dsCCPhosClient::Meta.Tables$TableName.Raw,
  CuratedTableNames = dsCCPhosClient::Meta.Tables$TableName.Curated,
  RunAssignmentChecks = TRUE,
  DSConnections = NULL
)
}
\arguments{
\item{ServerSpecifications}{\code{data.frame} - Same \code{data.frame} used for login. Used here only for acquisition of server-specific project names (in case they are differing) - Default: \code{NULL} for virtual project}

\item{RawTableNames}{\code{character vector} - The expected names of the Opal data base tables}

\item{CuratedTableNames}{\code{character vector} - The corresponding curated table names (used for assignment in R session)}

\item{RunAssignmentChecks}{\code{logical} Indicating whether assignment checks should be performed or omitted for reduced execution time - Default: \code{TRUE}}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of messages
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Load raw data set from Opal data base into R session on servers.
}
\author{
Bastian Reiter
}
