% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P21.CheckServerRequirements.R
\name{P21.CheckServerRequirements}
\alias{P21.CheckServerRequirements}
\title{P21.CheckServerRequirements}
\usage{
P21.CheckServerRequirements(
  ServerSpecifications = NULL,
  RequiredPackages = c("dsBase", "dsFredaP21", "dsTidyverse"),
  RequiredFunctions = c(assign = "P21.AugmentDataDS", assign = "P21.CurateDataDS", assign
    = "P21.DrawSampleDS", assign = "P21.PrepareRawDataSetDS"),
  DSConnections = NULL
)
}
\arguments{
\item{ServerSpecifications}{\code{data.frame} - Same \code{data.frame} used for login. Used here only for acquisition of server-specific project names (in case they are differing). - Default: \code{NULL} for virtual project}

\item{RequiredPackages}{A \code{character vector} naming required packages}

\item{RequiredFunctions}{A named \code{character vector} containing names of required functions. Their type ('aggregate' or 'assign') is defined by the correspondent element names.}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of \code{data.frames} containing gathered info and messages
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Check if technical requirements are met on every participating server.
}
\author{
Bastian Reiter
}
