% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{DisplayTimeSeries}
\alias{ds.ExtractFromList}
\alias{ds.FilterTable}
\alias{ds.GetCrossTab}
\alias{ds.GetCurationReport}
\alias{ds.GetDataSetCheck}
\alias{ds.GetFeatureInfo}
\alias{ds.GetFrequencyTable}
\alias{ds.GetObjectMetaData}
\alias{ds.GetObjectStatus}
\alias{ds.GetSampleStatistics}
\alias{ds.GetTableCheck}
\alias{ds.GetTTEModel}
\alias{ds.JoinTables}
\alias{ds.MakeList}
\alias{ds.MutateTable}
\alias{ExploreFeature}
\alias{ExportPlot}
\alias{GetEligibleValues}
\alias{GetServerOpalInfo}
\alias{GetServerWorkspaceInfo}
\alias{ggTheme}
\alias{gtTheme}
\alias{MakeBoxPlot}
\alias{MakeColumnPlot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dsFredaClient}{\code{\link[dsFredaClient]{DisplayTimeSeries}}, \code{\link[dsFredaClient]{ExploreFeature}}, \code{\link[dsFredaClient]{ExportPlot}}, \code{\link[dsFredaClient]{GetEligibleValues}}, \code{\link[dsFredaClient]{GetServerOpalInfo}}, \code{\link[dsFredaClient]{GetServerWorkspaceInfo}}, \code{\link[dsFredaClient]{MakeBoxPlot}}, \code{\link[dsFredaClient]{MakeColumnPlot}}, \code{\link[dsFredaClient]{ds.ExtractFromList}}, \code{\link[dsFredaClient]{ds.FilterTable}}, \code{\link[dsFredaClient]{ds.GetCrossTab}}, \code{\link[dsFredaClient]{ds.GetCurationReport}}, \code{\link[dsFredaClient]{ds.GetDataSetCheck}}, \code{\link[dsFredaClient]{ds.GetFeatureInfo}}, \code{\link[dsFredaClient]{ds.GetFrequencyTable}}, \code{\link[dsFredaClient]{ds.GetObjectMetaData}}, \code{\link[dsFredaClient]{ds.GetObjectStatus}}, \code{\link[dsFredaClient]{ds.GetSampleStatistics}}, \code{\link[dsFredaClient]{ds.GetTTEModel}}, \code{\link[dsFredaClient]{ds.GetTableCheck}}, \code{\link[dsFredaClient]{ds.JoinTables}}, \code{\link[dsFredaClient]{ds.MakeList}}, \code{\link[dsFredaClient]{ds.MutateTable}}, \code{\link[dsFredaClient]{ggTheme}}, \code{\link[dsFredaClient]{gtTheme}}}
}}

