% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.P21.DrawSample.R
\name{ds.P21.DrawSample}
\alias{ds.P21.DrawSample}
\title{ds.P21.DrawSample}
\usage{
ds.P21.DrawSample(
  RawDataSetName = "P21.RawDataSet",
  SampleSize = 100,
  SampleName = "P21.RDSSample",
  DSConnections = NULL
)
}
\arguments{
\item{RawDataSetName}{\code{string} - Name of a \code{list} object on server}

\item{SampleSize}{\code{integer} - Number of patients per Server}

\item{SampleName}{\code{string} - Option to assign subset of 'RawDataSet' a different object name on servers}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of messages about object assignment for monitoring purposes
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Draw a random sample from Raw Data Set
}
\details{
Linked to server-side ASSIGN method \code{DrawSampleDS()}
}
\author{
Bastian Reiter
}
