% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.P21.UnpackCuratedDataSet.R
\name{ds.P21.UnpackCuratedDataSet}
\alias{ds.P21.UnpackCuratedDataSet}
\title{ds.P21.UnpackCuratedDataSet}
\usage{
ds.P21.UnpackCuratedDataSet(
  CuratedDataSetName = "P21.CuratedDataSet",
  DSConnections = NULL
)
}
\arguments{
\item{CuratedDataSetName}{\code{string} - Name of Curated Data Set object (list) on server - Default: 'CuratedDataSet'}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}
}
\value{
A \code{list} of messages about object assignment for monitoring purposes
}
\description{
Make tables within Curated Data Set (\code{list} object) directly addressable in R server sessions
}
\details{
Linked to server-side \code{ASSIGN} function \code{ExtractFromListDS()}
}
\author{
Bastian Reiter
}
