% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P21.LoadRawDataSet.R
\name{P21.LoadRawDataSet}
\alias{P21.LoadRawDataSet}
\title{P21.LoadRawDataSet}
\usage{
P21.LoadRawDataSet(
  ServerSpecifications = NULL,
  ResourceNames.Mapping = setNames(dsFredaP21Client::Meta.Tables$TableName.Curated, nm =
    dsFredaP21Client::Meta.Tables$TableName.Raw),
  ResourceNames.Dictionary = NULL,
  RunAssignmentChecks = TRUE,
  DSConnections = NULL,
  DS.async = dsFredaClient::Set.DSSettings$DS.async
)
}
\arguments{
\item{ServerSpecifications}{\code{data.frame} - Same \code{data.frame} used for login. Used here only for acquisition of server-specific project names (in case they are differing) - Default: \code{NULL} for virtual project}

\item{ResourceNames.Mapping}{\code{character vector} - Usually the expected/required (generic) \code{resource} names. Can optionally be a named vector, with the names being expected names of \code{resources} and the vector values being the names of the assigned R \code{data.frames}.}

\item{ResourceNames.Dictionary}{Optional \code{list} of named \code{character vectors} - To enable server-specific mapping of deviating to required resource names. Names of list elements must match server names. For rules that should be applied on all servers, choose form \code{list(All = c('LookupName' = 'RequiredName'))}.}

\item{RunAssignmentChecks}{\code{logical} Indicating whether assignment checks should be performed or omitted for reduced execution time - Default: \code{TRUE}}

\item{DSConnections}{\code{list} of \code{DSConnection} objects. This argument may be omitted if such an object is already uniquely specified in the global environment.}

\item{DS.async}{\code{logical} - Value of argument 'async' in \code{DSI::datashield.assign()} / \code{DSI::datashield.aggregate()} - Default: \code{dsFredaClient::Set.DSSettings$DS.async}}
}
\value{
A \code{list} of messages
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Load data from resources into R session on servers.
}
\author{
Bastian Reiter
}
